+++ CelestiumPT/src/acceleration_structure/BLASBuilder.cu
@@ -9,7 +9,7 @@ void BLASBuilder::build(HostScene* hscene)
 {
 	DeviceScene* dscene = hscene->m_DeviceScene;
 	std::vector<BVHNode>hnodes;
-	std::vector<int>prim_indices;
+	std::vector<size_t>prim_indices;
 	const thrust::universal_vector<Triangle>& read_prims = dscene->DeviceTriangles;
 
 	BLAS::BVHBuilderSettings cfg;
